updating helper to avoid prettier errors
