+++ src/domain/deliver/deliver.js
@@ -1,5 +1,5 @@
 import { writeHTML, generateTemplateName } from '../write';
-import { printMessage } from '../../helper';
+import { printMessage } from '../helper-methods/cli';
 
 export const deliver = (template, name, message) => {
   const fileName = generateTemplateName(name);
