+++ src/domain/state/index.js
@@ -1,10 +1,6 @@
-//------------------
-// ---- helper > command-line-methods
-import {
-  checkErrors,
-  checkWarnings,
-  // checkHtml,
-} from '../../helper';
+import { checkErrors } from '../helper-methods/cli';
+import { checkWarnings } from '../helper-methods';
+
 // TODO convert into singleton
 const stateInit = (markdown, configurationMap) => {
   const stateObject = {
