+++ src/helper/index.js
@@ -5,25 +5,10 @@ import {
   generateTemplateName,
 } from '../domain/write';
 
-import {
-  displayCLIErrors,
-  checkErrors,
-  checkWarnings,
-  checkHtml,
-  printMessage,
-
-  // stateInit,
-} from './command-line-methods';
-
 export {
   writeHTML,
   readSourceFile,
   isFolderExists,
   generateTemplateName,
-  displayCLIErrors,
-  checkErrors,
-  checkWarnings,
-  checkHtml,
-  printMessage,
   // stateInit,
 };
